incus-osd/update: Move update logic to its own package
