+++ incus-osd/internal/update/doc.go
@@ -1,0 +1,2 @@
+// Package update contains logic for performing Secure Boot, OS, and application updates.
+package update
